Check CuratorFramework state before operating on it.

+++ extensions/spring-xd-extension-kafka/src/main/java/org/springframework/integration/x/kafka/KafkaPartitionAllocator.java
@@ -17,6 +17,8 @@
 
 package org.springframework.integration.x.kafka;
 
+import static org.apache.curator.framework.imps.CuratorFrameworkState.STARTED;
+
 import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -30,10 +32,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.curator.framework.CuratorFramework;
+import org.apache.curator.framework.imps.CuratorFrameworkState;
 import org.apache.curator.framework.recipes.locks.InterProcessMutex;
 import org.apache.curator.utils.EnsurePath;
 import org.apache.zookeeper.KeeperException;
 
+import org.springframework.beans.factory.BeanInitializationException;
 import org.springframework.beans.factory.FactoryBean;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.context.ApplicationListener;
@@ -99,8 +103,13 @@ public class KafkaPartitionAllocator implements InitializingBean, FactoryBean<Pa
 	public void afterPropertiesSet() throws Exception {
 		partitionDataPath = getDataPath();
 		EnsurePath ensurePath = new EnsurePath(partitionDataPath);
-		ensurePath.ensure(client.getZookeeperClient());
-		partitionDataMutex = new InterProcessMutex(client, partitionDataPath);
+		if (STARTED.equals(client.getState())) {
+			ensurePath.ensure(client.getZookeeperClient());
+			partitionDataMutex = new InterProcessMutex(client, partitionDataPath);
+		}
+		else {
+			throw new BeanInitializationException("Cannot connect to ZooKeeper, client state is " + client.getState());
+		}
 	}
 
 	@Override
@@ -112,71 +121,76 @@ public class KafkaPartitionAllocator implements InitializingBean, FactoryBean<Pa
 			log.debug("Sequence is " + sequence);
 			log.debug("Client is " + client);
 		}
-		try {
-			partitionDataMutex.acquire();
-			byte[] partitionData = client.getData().forPath(partitionDataPath);
-			if (partitionData == null || partitionData.length == 0) {
-				Collection<Partition> partitions = connectionFactory.getPartitions(topic);
-				final Map<Partition, BrokerAddress> leaders = connectionFactory.getLeaders(partitions);
-				ArrayList<Partition> sortedPartitions = new ArrayList<Partition>(partitions);
-				Collections.sort(sortedPartitions, new Comparator<Partition>() {
-					@Override
-					public int compare(Partition o1, Partition o2) {
-						int i = leaders.get(o1).toString().compareTo(leaders.get(o2).toString());
-						if (i != 0) {
-							return i;
+		if (STARTED.equals(client.getState())) {
+			try {
+				partitionDataMutex.acquire();
+				byte[] partitionData = client.getData().forPath(partitionDataPath);
+				if (partitionData == null || partitionData.length == 0) {
+					Collection<Partition> partitions = connectionFactory.getPartitions(topic);
+					final Map<Partition, BrokerAddress> leaders = connectionFactory.getLeaders(partitions);
+					ArrayList<Partition> sortedPartitions = new ArrayList<Partition>(partitions);
+					Collections.sort(sortedPartitions, new Comparator<Partition>() {
+						@Override
+						public int compare(Partition o1, Partition o2) {
+							int i = leaders.get(o1).toString().compareTo(leaders.get(o2).toString());
+							if (i != 0) {
+								return i;
+							}
+							else return o1.getId() - o2.getId();
 						}
-						else return o1.getId() - o2.getId();
-					}
-				});
-				if (log.isDebugEnabled()) {
-					log.debug("Partitions: " + StringUtils.collectionToCommaDelimitedString(sortedPartitions));
-				}
-				// calculate the minimum size of a partition group.
-				int minimumSize = sortedPartitions.size() / count;
-				int remainder = sortedPartitions.size() % count;
-				List<List<Integer>> partitionGroups = new ArrayList<List<Integer>>();
-				int cursor = 0;
-				for (int i = 0; i < count; i++) {
-					// first partitions will get an extra element
-					int partitionGroupSize = i < remainder ? minimumSize + 1 : minimumSize;
-					ArrayList<Integer> partitionGroup = new ArrayList<Integer>();
-					for (int j = 0; j < partitionGroupSize; j++) {
-						partitionGroup.add(sortedPartitions.get(cursor++).getId());
-					}
+					});
 					if (log.isDebugEnabled()) {
-						log.debug("Partitions for " + (i + 1) + " : " + StringUtils.collectionToCommaDelimitedString(partitionGroup));
+						log.debug("Partitions: " + StringUtils.collectionToCommaDelimitedString(sortedPartitions));
 					}
-					partitionGroups.add(partitionGroup);
+					// calculate the minimum size of a partition group.
+					int minimumSize = sortedPartitions.size() / count;
+					int remainder = sortedPartitions.size() % count;
+					List<List<Integer>> partitionGroups = new ArrayList<List<Integer>>();
+					int cursor = 0;
+					for (int i = 0; i < count; i++) {
+						// first partitions will get an extra element
+						int partitionGroupSize = i < remainder ? minimumSize + 1 : minimumSize;
+						ArrayList<Integer> partitionGroup = new ArrayList<Integer>();
+						for (int j = 0; j < partitionGroupSize; j++) {
+							partitionGroup.add(sortedPartitions.get(cursor++).getId());
+						}
+						if (log.isDebugEnabled()) {
+							log.debug("Partitions for " + (i + 1) + " : " + StringUtils.collectionToCommaDelimitedString(partitionGroup));
+						}
+						partitionGroups.add(partitionGroup);
+					}
+					byte[] dataAsBytes = objectMapper.writer().writeValueAsBytes(partitionGroups);
+					if (log.isDebugEnabled()) {
+						log.debug(new String(dataAsBytes));
+					}
+					client.setData().forPath(partitionDataPath, dataAsBytes);
+					// the partition mapping is stored 0-based but sequence/count are 1-based
+					if (log.isDebugEnabled()) {
+						log.debug("Listening to: " + StringUtils.collectionToCommaDelimitedString(partitionGroups.get(sequence - 1)));
+					}
+					return toPartitions(partitionGroups.get(sequence - 1));
 				}
-				byte[] dataAsBytes = objectMapper.writer().writeValueAsBytes(partitionGroups);
-				if (log.isDebugEnabled()) {
-					log.debug(new String(dataAsBytes));
+				else {
+					if (log.isDebugEnabled()) {
+						log.debug(new String(partitionData));
+					}
+					@SuppressWarnings("unchecked")
+					List<List<Integer>> partitionGroups = objectMapper.reader(List.class).readValue(partitionData);
+					// the partition mapping is stored 0-based but sequence/count are 1-based
+					if (log.isDebugEnabled()) {
+						log.debug("Listening to: " + StringUtils.collectionToCommaDelimitedString(partitionGroups.get(sequence - 1)));
+					}
+					return toPartitions(partitionGroups.get(sequence - 1));
 				}
-				client.setData().forPath(partitionDataPath, dataAsBytes);
-				// the partition mapping is stored 0-based but sequence/count are 1-based
-				if (log.isDebugEnabled()) {
-					log.debug("Listening to: " + StringUtils.collectionToCommaDelimitedString(partitionGroups.get(sequence - 1)));
-				}
-				return toPartitions(partitionGroups.get(sequence-1));
 			}
-			else {
-				if (log.isDebugEnabled()) {
-					log.debug(new String(partitionData));
+			finally {
+				if (partitionDataMutex.isAcquiredInThisProcess()) {
+					partitionDataMutex.release();
 				}
-				@SuppressWarnings("unchecked")
-				List<List<Integer>> partitionGroups = objectMapper.reader(List.class).readValue(partitionData);
-				// the partition mapping is stored 0-based but sequence/count are 1-based
-				if (log.isDebugEnabled()) {
-					log.debug("Listening to: " + StringUtils.collectionToCommaDelimitedString(partitionGroups.get(sequence - 1)));
-				}
-				return toPartitions(partitionGroups.get(sequence - 1));
 			}
 		}
-		finally {
-			if (partitionDataMutex.isAcquiredInThisProcess()) {
-				partitionDataMutex.release();
-			}
+		else {
+			throw new BeanInitializationException("Cannot connect to ZooKeeper, client state is " + client.getState());
 		}
 	}
 
@@ -204,44 +218,50 @@ public class KafkaPartitionAllocator implements InitializingBean, FactoryBean<Pa
 
 	@Override
 	public void onApplicationEvent(ContextClosedEvent event) {
-		try {
-			byte[] streamStatusData;
+		CuratorFrameworkState state = client.getState();
+		if (state.equals(STARTED)) {
 			try {
-				streamStatusData = client.getData().forPath(Paths.build(Paths.STREAM_DEPLOYMENTS, streamName, Paths.STATUS));
-			}
-			catch (KeeperException.NoNodeException e) {
-				// we ignore this error - the stream path does not exist, so it may have been removed already
-				// we'll behave as if we have received no data
-				streamStatusData = null;
-			}
-			DeploymentUnitStatus status = streamStatusData != null ?
-					new DeploymentUnitStatus(ZooKeeperUtils.bytesToMap(streamStatusData)) :
-					new DeploymentUnitStatus(DeploymentUnitStatus.State.undeployed);
-			if (DeploymentUnitStatus.State.undeployed.equals(status.getState())
-					|| DeploymentUnitStatus.State.undeploying.equals(status.getState())) {
-				if (client.checkExists().forPath(getDataPath()) != null) {
-					try {
-						client.delete().deletingChildrenIfNeeded().forPath(getDataPath());
-					}
-					catch (KeeperException.NoNodeException e) {
-						if (KeeperException.Code.NONODE.equals(e.code())) {
-							// ignore, most likely someone else has deleted the path already
+				byte[] streamStatusData;
+				try {
+					streamStatusData = client.getData().forPath(Paths.build(Paths.STREAM_DEPLOYMENTS, streamName, Paths.STATUS));
+				}
+				catch (KeeperException.NoNodeException e) {
+					// we ignore this error - the stream path does not exist, so it may have been removed already
+					// we'll behave as if we have received no data
+					streamStatusData = null;
+				}
+				DeploymentUnitStatus status = streamStatusData != null ?
+						new DeploymentUnitStatus(ZooKeeperUtils.bytesToMap(streamStatusData)) :
+						new DeploymentUnitStatus(DeploymentUnitStatus.State.undeployed);
+				if (DeploymentUnitStatus.State.undeployed.equals(status.getState())
+						|| DeploymentUnitStatus.State.undeploying.equals(status.getState())) {
+					if (client.checkExists().forPath(getDataPath()) != null) {
+						try {
+							client.delete().deletingChildrenIfNeeded().forPath(getDataPath());
 						}
-						else {
-							// this actually was an exception : we cannot recover, but cannot stop either
-							log.error(e);
-							e.printStackTrace();
+						catch (KeeperException.NoNodeException e) {
+							if (KeeperException.Code.NONODE.equals(e.code())) {
+								// ignore, most likely someone else has deleted the path already
+							}
+							else {
+								// this actually was an exception : we cannot recover, but cannot stop either
+								log.error(e);
+								e.printStackTrace();
+							}
 						}
 					}
 				}
+				else {
+					System.out.println("We cannot clean, because: " + status);
+				}
 			}
-			else {
-				System.out.println("We cannot clean, because: " + status);
+			catch (Exception e) {
+				log.error(e);
+				e.printStackTrace();
 			}
 		}
-		catch (Exception e) {
-			log.error(e);
-			e.printStackTrace();
+		else {
+			log.warn("Could not check the stream state and perform cleanup, client state was " + state);
 		}
 	}
 }
